requirements.txt
# Deep Learning Framework
torch==2.0.0
torchvision==0.15.0
torchaudio==2.0.0

# Data Processing & Scientific Computing
numpy==1.24.0
pandas==2.0.0
scipy==1.10.0
scikit-learn==1.3.0

# Visualization
matplotlib==3.7.0
seaborn==0.12.0
pillow==9.5.0

# Jupyter & Notebooks
jupyter==1.0.0
ipython==8.12.0
ipykernel==6.22.0

# Configuration & Utilities
pyyaml==6.0
tqdm==4.65.0
python-dotenv==1.0.0

# Adversarial Attack Libraries (Optional but Recommended)
advertorch==0.2.4
# AutoAttack - install from GitHub if using:
# pip install git+https://github.com/RobustBench/AutoAttack.git

# Debugging & Code Quality
black==23.3.0
flake8==6.0.0
isort==5.12.0

# Optional: For advanced visualization
plotly==5.13.0
tensorboard==2.12.0
